+++ src/aiomql/__init__.py
@@ -6,6 +6,7 @@ from .strategy import Strategy
 from .bot_builder import Bot
 from .result import Result
 from .records import Records
+from .trade_records import TradeRecords
 from .candle import Candle, Candles
 from .positions import Positions
 from .executor import Executor